Added some functionality to report offset
The detector scans every pixel in the block, and outputs the pixels that
can be identified.
The watcher picks these up and stores them in *.ok files.

The next step is to collect more information about the offsets and hard-
code them into detect.cpp ?

Perhaps we need to write a set union handler to get the common offset.