Cooked up ver.1 of screen number reader
It checks many times before deciding that a block is blank.
Also, the color of 8 is used in other stranged places, so we decided
to drop support for that.
But currently the app is dog slow, taking 92s to scan a mid-game.

Next step is to try using another way to read colors.