Pulled in finally.h and improved detect
Now we first copy the screen into a memory DC and read from that DC.
It's much faster. Reading a field with more zeroes took 16s, while the
previous version took about 92s. But that's still not fast enough.

Also, when writing mixed C++/WinAPI code, we need to be really careful
about resource leaks. So I brought in the finally.h to perform cleanup
actions that might not be done if exceptions are thrown.

Next step is to learn about the bitmaps a bit more and try to figure
out how to make it faster. I've seen on the Internet some programs don't
waste much time on reading the screen, and that's Python