Changed to more precise data and fixed a crit bug
1. x_step & y_step now are updated to more precise values.
    When a block is not opened, it has some interesting padding around
it, making it larger than it actually is. So the measurements must be
taken when the left up corner and the right down corner of the field are
both opened. I measured the coordinates carefully with 300% magnifier
and slowest possible mouse speed.
    However, additional data and repetitive research into the block width
height are still very much needed.

2. Critical bug in Holy::Butterfly::get_read_point()

  I tried to use sizeof to get the number of elements in a built-in arr,
which is obviously incorrect. I switched to std::array.
  C++ 17 is doing a great job in deducing the template arguments.