Introduced move idiom to dfs()
But this didn't prove so efficient as I thought. Instrumentation of the
previous and current version showed little difference in the execution
time of dfs(). In fact, the previous version was faster.

Next step is to try writing felix and roundup with the "frontier" con-
cept introduced, so they can search with much more specificity, which
might be faster.