Fixed bug and restored "testing" modifications
butterfly: changed seed of mt19937 from 0 to current time

deter_bench: 1. Added a string of w/l to indicate which games are won
and which games are lost.
2. Changed cycle to 100 (ancient bug)

Mineutils: Added checks for mines_left in mark_mine_check.

After this commit, the win rate of deter_bench has superceded that of
the BFS version (why?)

Next we will focus on the guesswork.